Aggregator/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gevent==20.9.0
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
requests==2.24.0
urllib3==1.25.11
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.1.2
python-dotenv